Refactor - subdirectories for related components
