+++ frontends/vue-webpack/src/components/shared/tripstable.vue
